Enum cgol::Cell

#[repr(u8)]pub enum Cell {
    Dead,
    Alive,
}

A "cell" which can be alive or dead.

It can be cast into any integer type using the as keyword. Alive becomes 1, and Dead becomes 0.

Internally, it is represented as a u8.

Notable Implementations

Using the unary operator ! will invert the cell.

Its default is Dead.

From and into bool.

From all integer types. Returns Alive if > 0, or Dead otherwise.

Into all integer types. Works the same as a cast.

Every mathematical operator for every integer type, as long as the integer is on the left. The cell will be cast before the operation.

Variants

Dead

Alive