Struct cgol::game::Game

pub struct Game { /* fields omitted */ }

An instance of Conway's Game of Life.

TODO: docs

Implementations

impl Game

pub fn new(width: usize, height: usize) -> Game

Creates a new instance of Conway's Game of Life.

Panics

Panics if width < 1, height < 1, or if width * height > isize::MAX.

pub fn tick(&mut self)

Ticks once.

Examples

let mut game = Game::new(100, 100);
game.fill_random(0.5);
game.tick();
// view the game somehow...

pub fn width(&self) -> usize

Gets this game's width.

Examples

let game = Game::new(69, 420);
assert_eq!(game.width(), 69);

pub fn height(&self) -> usize

Gets this game's height.

Examples

let game = Game::new(69, 420);
assert_eq!(game.height(), 420);

pub fn area(&self) -> usize

Gets this game's area.

Examples

let game = Game::new(4, 3);
assert_eq!(game.area(), 12);

pub fn get(&self, row: usize, col: usize) -> Option<Cell>

Gets a specific cell, returning None if out of bounds.

Example

let game = Game::new(10, 10);

assert_eq!(game.get(2, 3), Some(Dead));
assert_eq!(game.get(69, 420), None);

pub fn get_mut(&mut self, row: usize, col: usize) -> Option<&mut Cell>

Mutably gets a specific cell, returning None if out of bounds.

Example

let mut game = Game::new(3, 3);

if let Some(cell) = game.get_mut(0, 0) {
    *cell = Alive;
}

assert_eq!(game.get_row(0), &[Alive, Dead, Dead]);

pub fn get_row(&self, row: usize) -> &[Cell]

Gets a slice of a specific row.

Due to how cells are stored, this isn't possible for columns.

Panics

Panics if row is out of bounds.

Examples

let mut game = Game::new(3, 5);
assert_eq!(game.get_row(0), &[Dead, Dead, Dead]);

pub fn get_row_mut(&mut self, row: usize) -> &mut [Cell]

Gets a mutable slice of a specific row.

Due to how cells are stored, this isn't possible for columns.

Panics

Panics if row is out of bounds.

Examples

let mut game = Game::new(3, 5);
let mut row = game.get_row_mut(0);
row[0] = Alive;

assert_eq!(game.get_row(0), &[Alive, Dead, Dead]);

pub fn row(&self, row: usize) -> Row<'_>

Gets an iterator over the cells in a specific row.

Panics

Panics if row is out of bounds.

Examples

let game = Game::new(3, 5);
let mut row = game.row(0);

assert_eq!(row.next(), Some(Dead));
assert_eq!(row.next(), Some(Dead));
assert_eq!(row.next(), Some(Dead));
assert_eq!(row.next(), None);

pub fn row_mut(&mut self, row: usize) -> RowMut<'_>

Gets a mutable iterator over the cells in a specific row.

Panics

Panics if row is out of bounds.

Examples

let mut game = Game::new(3, 5);

for cell in game.row_mut(0) {
    *cell = Alive;
}

assert_eq!(game.get_row(0), &[Alive, Alive, Alive]);

pub fn col(&self, col: usize) -> Col<'_>

Gets an iterator over the cells in a specific column.

Panics

Panics if col is out of bounds.

Examples

let game = Game::new(7, 3);
let mut col = game.col(0);

assert_eq!(col.next(), Some(Dead));
assert_eq!(col.next(), Some(Dead));
assert_eq!(col.next(), Some(Dead));
assert_eq!(col.next(), None);

pub fn col_mut(&mut self, col: usize) -> ColMut<'_>

Gets a mutable iterator over the cells in a specific column.

Panics

Panics if col is out of bounds.

Examples

let mut game = Game::new(7, 3);

for cell in game.col_mut(0) {
    *cell = Alive;
}

let mut col: Vec<_> = game.col(0).collect();
assert_eq!(&col, &[Alive, Alive, Alive]);

pub fn clear(&mut self)

Kills all cells.

Examples

let mut game = Game::new(10, 10);
game[(0, 0)] = Alive;
game.clear();
assert_eq!(game[(0, 0)], Dead);

pub fn all_dead(&self) -> bool

Returns true if all cells are Dead.

Examples

let mut game = Game::new(10, 10);
assert!(game.all_dead());

game[(0, 0)] = Alive;
assert!(!game.all_dead())

pub fn all_alive(&self) -> bool

Returns true if all cells are Alive.

Examples

let mut game = Game::new(10, 10);
assert!(!game.all_alive());

game.invert();
assert!(game.all_alive());

pub fn invert(&mut self)

Inverts all cells.

Examples

let mut game = Game::new(10, 10);
game.invert();
assert!(game.all_alive());

pub fn fill_random(&mut self, chance: f64)

Fills the game's cells randomly with a probability of being alive.

Panics

Panics if chance is not in the range [0, 1].

Examples

let mut game = Game::new(10, 10);

game.fill_random(1.0);
assert!(game.all_alive());

game.fill_random(0.0);
assert!(game.all_dead());

game.fill_random(0.5);
assert!(!game.all_dead());